+++ ray.h
@@ -11,7 +11,7 @@ public:
     const point3& origin() const { return orig; }
     const point3& direction() const { return dir; }
 
-    point3 at(double t) const {
+    point3 at(float t) const {
         return orig + t*dir;
     }
