+++ sphere.h
@@ -6,7 +6,7 @@
 
 class sphere : public hittable {
 public:
-  sphere(const point3 &_center, double _radius, shared_ptr<material> mat) : center(_center), radius(std::fmax(0, _radius)), mat(mat) {}
+  sphere(const point3 &_center, float _radius, shared_ptr<material> mat) : center(_center), radius(std::fmax(0, _radius)), mat(mat) {}
   bool hit(const ray &r, interval ray_t, hit_record &rec) const override {
     // The math behond it is the sphere equation of x^2 + y^2 + z^2 = r^2,
     // Which with 2 points just becomes (x1-x2)^2 and resp.for y and z,
@@ -56,7 +56,7 @@ public:
 
 private:
   point3 center;
-  double radius;
+  float radius;
   shared_ptr<material> mat;
 };
